requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==3.2.18
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-taggit==3.1.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ez-setup==0.9
Pillow==9.4.0
pkg_resources==0.0.0
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
